requirements.txt
# Core requirements
requests==2.31.0
beautifulsoup4==4.12.2
psycopg2-binary==2.9.9
python-dotenv==1.0.0

# For future extensions
SQLAlchemy==1.4.51
alembic==1.11.3